tests/wpt/tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-async-fetch-disconnect-iframe.html

<!DOCTYPE html>
<title>shadowrootadoptedstylesheets async fetch with iframe destruction</title>
<meta name="author" title="Kurt Catti-Schmidt" href="mailto:kschmi@microsoft.com" />
<link rel='help' href='https://github.com/whatwg/html/pull/11687'>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='./support/helpers.js'></script>
<body>
<script type="module">
  promise_test(async (t) => {
    // Host is inside an iframe that is destroyed during the fetch. Destroying
    // the iframe invalidates its browsing context, which causes the pending
    // module fetch to fail. The cleanup callback must handle this without
    // crashing; the orphaned shadow root retains its synchronously-adopted
    // empty sheet.
    const cssUrl = new URL("./support/styles-red.css?iframe", location.href).href;

    const iframe = document.createElement("iframe");
    const iframeLoaded = new Promise(
      resolve => iframe.addEventListener("load", resolve)
    );
    document.body.appendChild(iframe);
    await iframeLoaded;

    iframe.contentDocument.body.setHTMLUnsafe(
      "<div id='host'>" +
        "<template shadowrootmode='open' " +
          `shadowrootadoptedstylesheets='${cssUrl}'>` +
          "<span>Test content</span>" +
        "</template>" +
      "</div>"
    );

    const iframeHost = iframe.contentDocument.getElementById("host");
    const shadowRoot = iframeHost.shadowRoot;

    assert_equals(shadowRoot.adoptedStyleSheets.length, 1,
      "Before iframe destruction: one entry exists synchronously.");
    assert_equals(shadowRoot.adoptedStyleSheets[0].cssRules.length, 0,
      "Before iframe destruction: entry at index 0 is empty pre-fetch.");

    iframe.remove();

    await new Promise(resolve => step_timeout(resolve, 500));
    await new Promise(resolve => requestAnimationFrame(resolve));

    // After the browsing context is invalidated, the failed fetch is
    // observed without crashing. The empty sheet is left in place.
    assert_equals(shadowRoot.adoptedStyleSheets.length, 1,
      "After iframe destruction: empty sheet remains.");
    assert_equals(shadowRoot.adoptedStyleSheets[0].cssRules.length, 0,
      "After iframe destruction: sheet remains empty.");

    // Re-inserting the iframe creates a new browsing context with a fresh document.
    const iframeReloaded = new Promise(
      resolve => iframe.addEventListener("load", resolve)
    );
    document.body.appendChild(iframe);
    await iframeReloaded;

    assert_equals(iframe.contentDocument.getElementById("host"), null,
      "The new iframe document should not contain the original host element.");
    assert_equals(shadowRoot.adoptedStyleSheets.length, 1,
      "After iframe re-insertion: orphaned shadow root still holds sheet.");
  }, "Destroying iframe during async adoptedStyleSheets fetch does not crash.");
</script>
</body>
